refactor: rename manage run support package
