+++ src/infrastructure/dpm_runs/in_memory.py
@@ -4,7 +4,7 @@ from datetime import datetime, timedelta, timezone
 from threading import Lock
 from typing import Any, Optional
 
-from src.core.dpm_runs.models import (
+from src.core.rebalance_runs.models import (
     DpmAsyncOperationRecord,
     DpmLineageEdgeRecord,
     DpmRunIdempotencyHistoryRecord,
@@ -13,7 +13,7 @@ from src.core.dpm_runs.models import (
     DpmRunWorkflowDecisionRecord,
     DpmSupportabilitySummaryData,
 )
-from src.core.dpm_runs.repository import DpmRunRepository
+from src.core.rebalance_runs.repository import DpmRunRepository
 
 
 class InMemoryDpmRunRepository(DpmRunRepository):
