+++ tests/unit/dpm/supportability/test_dpm_idempotency_history_service.py
@@ -4,7 +4,7 @@ from decimal import Decimal
 import pytest
 
 from src.core.dpm.engine import run_simulation
-from src.core.dpm_runs.service import DpmRunNotFoundError, DpmRunSupportService
+from src.core.rebalance_runs.service import DpmRunNotFoundError, DpmRunSupportService
 from src.core.models import EngineOptions
 from src.infrastructure.dpm_runs import InMemoryDpmRunRepository
 from tests.shared.factories import (
